댓글이 없는 게시글 삭제 시 '댓글 실패' 메시지가 출력되는 오류 수정

+++ src/main/java/edu/kh/nndr/mainFeed/model/service/ReplyServiceImpl.java
@@ -29,6 +29,9 @@ public class ReplyServiceImpl implements ReplyService {
 	// 게시글 삭제 시 종속된 댓글 모두 삭제
 	@Override
 	public int deleteReplyAll(int boardNo) {
+		
+		// 게시글에 달린 댓글이 없는 경우 1(sql 정상수행 응답) 반환
+		if(dao.count(boardNo) == 0) return 1;
 		return dao.deleteAll(boardNo);
 	}
 
@@ -48,5 +51,11 @@ public class ReplyServiceImpl implements ReplyService {
 	@Override
 	public int update(Reply reply) {
 		return dao.update(reply);
-	}	
+	}
+	
+	// 특정 게시글 댓글 갯수 조회
+	@Override
+	public int count(int boardNo) {
+		return dao.count(boardNo);
+	}
 }
